feat: feat/add-sops
