+++ sops/dagger.json
@@ -1,0 +1,8 @@
+{
+  "name": "sops",
+  "engineVersion": "v0.18.10",
+  "sdk": {
+    "source": "go"
+  },
+  "source": "."
+}
