+++ tests/sops/tfvars.json
@@ -1,0 +1,5 @@
+{
+  "aws_access_key": "AKIAEXAMPLE",
+  "aws_secret_key": "abc123EXAMPLE",
+  "region": "eu-central-1"
+}
